# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Terramaps API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/maps": {
      "get": {
        "tags": [
          "Maps"
        ],
        "summary": "List Maps",
        "description": "List maps for the current user, each with its active job and import state.",
        "operationId": "list_maps_maps_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Map"
                  },
                  "type": "array",
                  "title": "Response List Maps Maps Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Maps"
        ],
        "summary": "Create Map",
        "description": "Create a map from a previously uploaded and parsed document.",
        "operationId": "create_map_maps_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMap"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Map"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/maps/{map_id}": {
      "patch": {
        "tags": [
          "Maps"
        ],
        "summary": "Patch Map",
        "description": "Rename a map. Any map member.",
        "operationId": "patch_map_maps__map_id__patch",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchMapDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Map"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Maps"
        ],
        "summary": "Get Map",
        "description": "Get a single map with its active job and import state.",
        "operationId": "get_map_maps__map_id__get",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Map"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/maps/{map_id}/export/ztt": {
      "get": {
        "tags": [
          "Exports"
        ],
        "summary": "Export Ztt",
        "description": "Export a map's zip-to-territory hierarchy as an Excel (.xlsx) file.\n\nColumns: zip_code | <layer order=1 name> | <layer order=2 name> | \u2026\nOne row per zip code; territory columns are blank for unassigned zips.",
        "operationId": "export_ztt_maps__map_id__export_ztt_get",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Map Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateMap": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "layers": {
            "items": {
              "$ref": "#/components/schemas/LayerSetup"
            },
            "type": "array",
            "title": "Layers"
          },
          "data_fields": {
            "items": {
              "$ref": "#/components/schemas/DataFieldSetup"
            },
            "type": "array",
            "title": "Data Fields"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "name",
          "layers",
          "data_fields"
        ],
        "title": "CreateMap",
        "description": "DTO for POST /maps.\n\nReferences a previously uploaded and parsed document. The raw spreadsheet\ndata stays in S3 and is read by the background import worker."
      },
      "DataFieldConfig": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "default": ""
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "number"
            ],
            "title": "Type"
          },
          "aggregations": {
            "items": {
              "type": "string",
              "enum": [
                "sum",
                "avg",
                "min",
                "max"
              ]
            },
            "type": "array",
            "title": "Aggregations"
          },
          "precision": {
            "type": "integer",
            "title": "Precision",
            "default": 4
          }
        },
        "type": "object",
        "required": [
          "field",
          "type",
          "aggregations"
        ],
        "title": "DataFieldConfig",
        "description": "Data field config entry stored on a map."
      },
      "DataFieldSetup": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "header": {
            "type": "string",
            "title": "Header"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "number"
            ],
            "title": "Type"
          },
          "aggregations": {
            "items": {
              "type": "string",
              "enum": [
                "sum",
                "avg",
                "min",
                "max"
              ]
            },
            "type": "array",
            "title": "Aggregations",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "header",
          "type"
        ],
        "title": "DataFieldSetup",
        "description": "DataFieldSetup."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LayerSetup": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "header": {
            "type": "string",
            "title": "Header"
          }
        },
        "type": "object",
        "required": [
          "name",
          "header"
        ],
        "title": "LayerSetup",
        "description": "LayerSetup."
      },
...
      "Map": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tile_version": {
            "type": "integer",
            "title": "Tile Version",
            "default": 0
          },
          "data_field_config": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DataFieldConfig"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Field Config"
          },
          "active_job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MapJob"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_export": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MapExport"
              },
              {
                "type": "null"
              }
            ]
          },
          "import_state": {
            "$ref": "#/components/schemas/MapImportState"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "import_state"
        ],
        "title": "Map",
        "description": "Map."
      },
      "MapExport": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "in_progress",
              "generating",
              "complete",
              "failed"
            ],
            "title": "Status"
          },
          "total_slides": {
            "type": "integer",
            "title": "Total Slides"
          },
          "uploaded_slides": {
            "type": "integer",
            "title": "Uploaded Slides"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "total_slides",
          "uploaded_slides"
        ],
        "title": "MapExport",
        "description": "Active PPT export session for a map."
      },
      "MapImportState": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "importing",
              "complete",
              "failed"
            ],
            "title": "Status"
          },
          "step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step"
          },
          "warnings": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warnings"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "MapImportState",
        "description": "Import lifecycle state embedded in Map responses.\n\nCovers the import phase (after POST /maps). The parse phase lives on MapUploadStatus.\nNon-nullable on Map \u2014 all maps in this system are created via the upload flow."
      },
...
      "MapJob": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "job_type": {
            "type": "string",
            "enum": [
              "recompute_geometry",
              "recompute_data"
            ],
            "title": "Job Type"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "complete",
              "failed"
            ],
            "title": "Status"
          },
          "step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_type",
          "status"
        ],
        "title": "MapJob",
        "description": "Background recompute job for a map (geometry or data). Import lifecycle is tracked separately via MapImportState."
      },
...
      "PatchMapDTO": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PatchMapDTO"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}